Change projection reset functionality
